Restructure tokens directories and themeing
