+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@linode/design-language-system",
-  "version": "0.0.11",
+  "version": "0.0.12",
   "private": false,
   "type": "module",
   "exports": {
