+++ tokens/global/highContrast.json
@@ -1,0 +1,1 @@
+{}
