+++ tsup.config.ts
@@ -3,8 +3,9 @@ import { defineConfig } from 'tsup';
 export default defineConfig({
   dts: true,
   entry: [
-    ...['config/akamai/index.ts', 'config/cloudmanager/index.ts'],
-    ...['config/akamai/dark/index.ts', 'config/cloudmanager/dark/index.ts']
+    ...['config/index.ts'],
+    ...['config/dark/index.ts'],
+    ...['config/highContast/index.ts']
   ],
   format: ['esm'],
   minify: 'terser',
